refactor mod output to use more files
