+++ src/output/file.rs
@@ -8,36 +8,8 @@ use std::time::Duration;
 
 use item::Item;
 
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub enum OutputErrorKind {
-    PrepareError(String),
-    WriteError(String),
-    CleanupError(String),
-}
-
-#[derive(Debug)]
-pub struct OutputError {
-    kind: OutputErrorKind,
-    cause: Option<Box<::std::error::Error>>,
-}
-
-impl ::std::fmt::Display for OutputError {
-    fn fmt(&self, f: &mut ::std::fmt::Formatter) -> Result<(), ::std::fmt::Error> {
-        match self.kind {
-            OutputErrorKind::PrepareError(ref s) => write!(f, "failed to prepare output {}", s),
-            OutputErrorKind::WriteError(ref s) => write!(f, "failed writing values to output {}", s),
-            OutputErrorKind::CleanupError(ref s) => write!(f, "cleanup of output {} returned an error", s)
-        }
-    }
-}
-
-pub trait AKOutput {
-    fn prepare(&mut self, items: &Vec<Item>) -> Result<(), OutputError>;
-    fn write_value(&mut self, key: &String, time: Duration, value: f64) -> Result<(), OutputError>;
-    fn write_raw_value(&mut self, key: &String, time: Duration, value: &String) -> Result<(), OutputError>;
-    fn write_raw_value_as_fallback(&mut self, key: &String, time: Duration, value: &String) -> Result<(), OutputError>;
-    fn clean_up(&mut self) -> Result<(), OutputError>;
-}
+use output::AKOutput;
+use output::error::*;
 
 #[derive(Debug, Clone, Eq, PartialEq, Deserialize)]
 pub struct FileOutput {
@@ -130,48 +102,5 @@ impl AKOutput for FileOutput {
 
     fn clean_up(&mut self) -> Result<(), OutputError> {
         Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Deserialize)]
-#[serde(tag = "type", rename_all = "lowercase")]
-pub enum OutputKind {
-    File{
-        #[serde(flatten)]
-        fo : FileOutput
-    },
-    // more in the future
-}
-
-impl AKOutput for OutputKind {
-
-    fn prepare(&mut self, items: &Vec<Item>) -> Result<(), OutputError> {
-        match self {
-            OutputKind::File{ fo } => fo.prepare(items),
-        }
-    }
-
-    fn write_value(&mut self, key: &String, time: Duration, value: f64) -> Result<(), OutputError> {
-        match self {
-            OutputKind::File{ fo } => fo.write_value(key, time, value),
-        }
-    }
-
-    fn write_raw_value_as_fallback(&mut self, key: &String, time: Duration, value: &String) -> Result<(), OutputError> {
-        match self {
-            OutputKind::File{ fo } => fo.write_raw_value_as_fallback(key, time, value),
-        }
-    }
-
-    fn write_raw_value(&mut self, key: &String, time: Duration, value: &String) -> Result<(), OutputError> {
-        match self {
-            OutputKind::File{ fo } => fo.write_raw_value(key, time, value),
-        }
-    }
-
-    fn clean_up(&mut self) -> Result<(), OutputError> {
-        match self {
-            OutputKind::File{ fo } => fo.clean_up(),
-        }
     }
 }
